Get MacOS Dasher building and working in 64-bit (#171)
* Update to 64-bit macOS build. Changes sponsored and supplied by Jorge Tendeiro

* Add Debug xcode scheme

* Fix crash when opening Preferences for 2nd time by using strong reference to panel. Remove unused instance variable.

* Fix incorrect permissions when first creating user's training file

* Fix string scoping bug which prevented user's training text being read on startup. Removed unused method.

* Revert DasherIntefaceBase::m_FileUtils to private and set it in the expected way through the constructor

* Whitespace

* Replace tab indents with spaces

* Whitespace to match original, so there are no diffs outside MacOSX folder

* Set ApplicationCategory metadata to the most appropriate of the limited options available

* Update README, add INSTALL.macOS

* Update gitignore

+++ Src/MacOSX/AppWatcher.h
@@ -20,8 +20,8 @@
 - (id)init;
 - (void)awakeFromNib;
 - (void)fillApps;
-- (unsigned int)indexOfAppWithApplicationName:(NSString *)aName;
-- (unsigned int)indexOfAppWithInfo:(NSDictionary *)anAppInfo;
+- (unsigned long)indexOfAppWithApplicationName:(NSString *)aName;
+- (unsigned long)indexOfAppWithInfo:(NSDictionary *)anAppInfo;
 - (void)seeAppLaunch:(NSNotification *)note;
 - (NSString *)applicationName;
 - (BOOL)addAppInfo:(NSDictionary *)anAppInfo;
